Made mods for modularity/privacy
